bot/requirements.txt
beautifulsoup4
discord.py
matplotlib
Pillow
python-dotenv
requests
yt-dlp
PyNaCl